Project_backend/requirements.txt
asgiref==3.11.1
boto3==1.42.66
botocore==1.42.66
azure-identity==1.19.0
azure-mgmt-resource==22.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-compute==1.46.0
google-cloud-core==2.5.0
google-cloud-iam==2.21.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
idna==3.11
jmespath==1.1.0
pillow==12.1.1
proto-plus==1.27.1
protobuf==6.33.6
psycopg2==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
reportlab==4.4.10
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
openpyxl==3.1.2
